requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.34.34
pandas==2.1.4
pyarrow==15.0.0
python-dotenv==1.0.0
numpy==1.26.4
# openpyxl est requis par pandas pour lire les .xlsx
openpyxl==3.1.2
# Dependencies pour les tests
pytest==7.4.3
moto==4.2.14
requests==2.31.0 